requirements.txt
redis
python-dotenv
redisvl
slack-sdk
langchain-openai
langchain-community
PyPDF2
nltk
# Optional enhanced PDF processing (install one of these for better quality)
# pdfplumber>=0.9.0  # Recommended for best PDF extraction
# pymupdf>=1.23.0    # Alternative PDF processor 